Move new options into "extraOptions"
The lloyd fork will crash if it sees unexpected option names in the
"options" key of a save file. This moves "notifyOnPause" to a new key,
"extraOptions", which older forks will ignore entirely.

This also fixes the crash on unexpected option names for this fork,
so that it can parse saves coming from other versions (potentially even
saves coming from newer versions of this fork).