Copy over the already-written stories
Credit to Inferno Vulpix, Nikki, and scout (me) for the stories.

I think Vera's stories are generally great, and some are amazing! They're not always perfect, though, so when there was a story by Vera in the same slot as a story from the spreadsheet, I tried to use whichever one made the stories better overall.

TODOs:
- I left all the ruins and haul actions alone for now, since we only have (non-placeholder) stories written for a few of those.
- Some of the training action stories could use further improvement.
- I love IV's Wizard College stories, but Vera's are good too. Need to figure out what to do about that.
- IV wrote a cool set of stories for Fell From Grace with more and more required resources.  The numbers probably need some tweaking.